refactor(analysis): split content complexity function tests
