+++ docs/architecture-consolidation-plan.md
@@ -59,7 +59,7 @@ fixtures:
 
 | Area | Current file | Approx. lines | Owner direction |
 | --- | --- | ---: | --- |
-| Content complexity | `crates/tokmd-analysis/src/content/complexity.rs` and `crates/tokmd-analysis/src/content/complexity/` | `tests/unit.rs` 1451; production owner modules <=187 | Scoring, nesting, and function-span helpers now live under owner modules; remaining work is mostly test split and aggregation cleanup |
+| Content complexity | `crates/tokmd-analysis/src/content/complexity.rs` and `crates/tokmd-analysis/src/content/complexity/` | `tests/unit.rs` 995; function test owner 461; production owner modules <=187 | Scoring, nesting, and function-span helpers now live under owner modules; function-detection proof tests now have a dedicated owner; remaining work is mostly cyclomatic/cognitive/nesting test split and aggregation cleanup |
 | Content I/O helpers | `crates/tokmd-analysis/src/content/io.rs` and `crates/tokmd-analysis/src/content/io/` | coordinator 351; tag owner/tests 71 | Keep file reads, text detection, hashing, and entropy helpers in `io.rs` for now; keep tag counting and delimited matching in `io/tags.rs`. Preserve `crate::content::io::*` call sites and the `analysis_content_assets` proof scope. |
 | Analysis orchestration | `crates/tokmd-analysis/src/analysis/mod.rs` and `crates/tokmd-analysis/src/analysis/` | coordinator 152; setup 20; files 53; outputs 22; enricher owners <=160 | Derived/source setup, file collection, optional output storage, and inventory/content/git/semantic/code-quality/effort enrichers now live in owner modules while `analyze` remains the public coordinator. Keep `analysis_orchestration` proof scoped to both the historical `analysis.rs` path and the new `analysis/**` owner paths during the file-to-folder transition |
 | Analysis git report | `crates/tokmd-analysis/src/git/mod.rs`, `git/freshness.rs`, and `git/churn.rs` | coordinator 221; freshness/code-age owner 192; predictive churn owner/tests 174 | Keep export-row matching, hotspot, bus-factor, coupling, and intent assembly in `mod.rs` for now; keep freshness, age buckets, and refresh trend in `freshness.rs`; keep predictive churn in `churn.rs`. Preserve the `analysis_git` proof scope. |
